requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-nested-admin==3.4.0
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.2
